fix(avatars): convert MediaObject upload to a state processor
CI's KernelBrowser-driven test path doesn't have API Platform 4's
kernel.view bridge in front of custom controllers, so the previous
controller-based upload returned an entity and Symfony rejected the
response with "controller must return a Response object". A state
processor stays inside API Platform's pipeline, so the returned
MediaObject is serialized normally.

Also fixes testRegisterUser by initializing User.personalizedColor to a
palette member (slate-700). The previous '#1e6091' default tripped the
new Assert\Choice validator on POST, before the password-hasher
processor had a chance to overwrite it with a random palette pick.
